Changed the test point classification, instead of reading it of a list in the code i now read in the file containing the testpoints.
Implemented a new classification algorithm for user
imput since the first one was flawed, sums can differ
depending on the closest point in the trainingdata

Now we use the closest 5 points and depening on
which pokemon has the majority we return that pokemon.
This need some tweaking to then print out and take the user data.

easy fix with np.coloum_stack since this combines the two
np.arrays to a 2D array. We can then use a simple if statement and counts
to count which one has the closest points in range(5)

next step:

clean code, correct indentation, do this manually, I will then tweak and
remove any excessive code,

ask teacher or ChatGPT to look over the OOP aspects of the code, maybe new class?